+++ src/lang-xquery/test/uk/co/reecedunn/intellij/plugin/xquery/tests/parser/XQueryParserTest.kt
@@ -9714,6 +9714,34 @@ private class XQueryParserTest : ParserTestCase() {
     }
 
     @Nested
+    @DisplayName("XQuery 4.0 ED EBNF (122) RecordTest")
+    internal inner class RecordTest {
+        @Test
+        @DisplayName("record test")
+        fun recordTest() {
+            val expected = loadResource("tests/parser/xquery-4.0/RecordTest.txt")
+            val actual = parseResource("tests/parser/xquery-4.0/RecordTest.xq")
+            assertThat(actual.toPsiTreeString(), `is`(expected))
+        }
+
+        @Test
+        @DisplayName("record test; compact whitespace")
+        fun recordTest_CompactWhitespace() {
+            val expected = loadResource("tests/parser/xquery-4.0/RecordTest_CompactWhitespace.txt")
+            val actual = parseResource("tests/parser/xquery-4.0/RecordTest_CompactWhitespace.xq")
+            assertThat(actual.toPsiTreeString(), `is`(expected))
+        }
+
+        @Test
+        @DisplayName("missing closing parenthesis")
+        fun missingClosingParenthesis() {
+            val expected = loadResource("tests/parser/xquery-4.0/RecordTest_MissingClosingParenthesis.txt")
+            val actual = parseResource("tests/parser/xquery-4.0/RecordTest_MissingClosingParenthesis.xq")
+            assertThat(actual.toPsiTreeString(), `is`(expected))
+        }
+    }
+
+    @Nested
     @DisplayName("XQuery Terminal Delimitation")
     internal inner class TerminalDelimitation {
         @Test